add GetS3Keys unit test
